+++ include/wlan_bssdef.h
@@ -32,11 +32,11 @@ typedef long    		NDIS_802_11_RSSI;           /* in dBm */
 typedef unsigned char   NDIS_802_11_RATES[NDIS_802_11_LENGTH_RATES];        /* Set of 8 data rates */
 typedef unsigned char   NDIS_802_11_RATES_EX[NDIS_802_11_LENGTH_RATES_EX];  /* Set of 16 data rates */
 
-typedef  ULONG  NDIS_802_11_KEY_INDEX;
+typedef  u32  NDIS_802_11_KEY_INDEX;
 typedef unsigned long long NDIS_802_11_KEY_RSC;
 
 typedef struct _NDIS_802_11_SSID {
-	ULONG  SsidLength;
+	u32  SsidLength;
 	u8  Ssid[32];
 } NDIS_802_11_SSID, *PNDIS_802_11_SSID;
 
@@ -49,10 +49,10 @@ typedef enum _NDIS_802_11_NETWORK_TYPE {
 } NDIS_802_11_NETWORK_TYPE, *PNDIS_802_11_NETWORK_TYPE;
 
 typedef struct _NDIS_802_11_CONFIGURATION_FH {
-	ULONG           Length;             /* Length of structure */
-	ULONG           HopPattern;         /* As defined by 802.11, MSB set */
-	ULONG           HopSet;             /* to one if non-802.11 */
-	ULONG           DwellTime;          /* units are Kusec */
+	u32           Length;             /* Length of structure */
+	u32           HopPattern;         /* As defined by 802.11, MSB set */
+	u32           HopSet;             /* to one if non-802.11 */
+	u32           DwellTime;          /* units are Kusec */
 } NDIS_802_11_CONFIGURATION_FH, *PNDIS_802_11_CONFIGURATION_FH;
 
 /*
@@ -60,10 +60,10 @@ typedef struct _NDIS_802_11_CONFIGURATION_FH {
 	ODI Handler will convert the channel number to freq. number.
 */
 typedef struct _NDIS_802_11_CONFIGURATION {
-	ULONG           Length;             /* Length of structure */
-	ULONG           BeaconPeriod;       /* units are Kusec */
-	ULONG           ATIMWindow;         /* units are Kusec */
-	ULONG           DSConfig;           /* channel number */
+	u32           Length;             /* Length of structure */
+	u32           BeaconPeriod;       /* units are Kusec */
+	u32           ATIMWindow;         /* units are Kusec */
+	u32           DSConfig;           /* channel number */
 	NDIS_802_11_CONFIGURATION_FH    FHConfig;
 } NDIS_802_11_CONFIGURATION, *PNDIS_802_11_CONFIGURATION;
 
@@ -91,7 +91,7 @@ typedef struct _NDIS_802_11_VARIABLE_IEs {
 /*
 
 Length is the 4 bytes multiples of the sume of
-	sizeof (NDIS_802_11_MAC_ADDRESS) + 2 + sizeof (NDIS_802_11_SSID) + sizeof (ULONG)
+	sizeof (NDIS_802_11_MAC_ADDRESS) + 2 + sizeof (NDIS_802_11_SSID) + sizeof (u32)
 +   sizeof (NDIS_802_11_RSSI) + sizeof (NDIS_802_11_NETWORK_TYPE) + sizeof (NDIS_802_11_CONFIGURATION)
 +   sizeof (NDIS_802_11_RATES_EX) + IELength
 
@@ -148,15 +148,15 @@ typedef struct _NDIS_802_11_AI_RESFI {
 } NDIS_802_11_AI_RESFI, *PNDIS_802_11_AI_RESFI;
 
 typedef struct _NDIS_802_11_ASSOCIATION_INFORMATION {
-	ULONG                   Length;
+	u32                   Length;
 	u16                  AvailableRequestFixedIEs;
 	NDIS_802_11_AI_REQFI    RequestFixedIEs;
-	ULONG                   RequestIELength;
-	ULONG                   OffsetRequestIEs;
+	u32                   RequestIELength;
+	u32                   OffsetRequestIEs;
 	u16                  AvailableResponseFixedIEs;
 	NDIS_802_11_AI_RESFI    ResponseFixedIEs;
-	ULONG                   ResponseIELength;
-	ULONG                   OffsetResponseIEs;
+	u32                   ResponseIELength;
+	u32                   OffsetResponseIEs;
 } NDIS_802_11_ASSOCIATION_INFORMATION, *PNDIS_802_11_ASSOCIATION_INFORMATION;
 
 typedef enum _NDIS_802_11_RELOAD_DEFAULTS {
@@ -166,31 +166,31 @@ typedef enum _NDIS_802_11_RELOAD_DEFAULTS {
 
 /* Key mapping keys require a BSSID */
 typedef struct _NDIS_802_11_KEY {
-	ULONG           Length;             /* Length of this structure */
-	ULONG           KeyIndex;
-	ULONG           KeyLength;          /* length of key in bytes */
+	u32           Length;             /* Length of this structure */
+	u32           KeyIndex;
+	u32           KeyLength;          /* length of key in bytes */
 	NDIS_802_11_MAC_ADDRESS BSSID;
 	NDIS_802_11_KEY_RSC KeyRSC;
 	u8           KeyMaterial[32];     /* variable length depending on above field */
 } NDIS_802_11_KEY, *PNDIS_802_11_KEY;
 
 typedef struct _NDIS_802_11_REMOVE_KEY {
-	ULONG                   Length;        /* Length of this structure */
-	ULONG                   KeyIndex;
+	u32                   Length;        /* Length of this structure */
+	u32                   KeyIndex;
 	NDIS_802_11_MAC_ADDRESS BSSID;
 } NDIS_802_11_REMOVE_KEY, *PNDIS_802_11_REMOVE_KEY;
 
 typedef struct _NDIS_802_11_WEP {
-	ULONG     Length;        /* Length of this structure */
-	ULONG     KeyIndex;      /* 0 is the per-client key, 1-N are the global keys */
-	ULONG     KeyLength;     /* length of key in bytes */
+	u32     Length;        /* Length of this structure */
+	u32     KeyIndex;      /* 0 is the per-client key, 1-N are the global keys */
+	u32     KeyLength;     /* length of key in bytes */
 	u8     KeyMaterial[16];/* variable length depending on above field */
 } NDIS_802_11_WEP, *PNDIS_802_11_WEP;
 
 typedef struct _NDIS_802_11_AUTHENTICATION_REQUEST {
-	ULONG Length;            /* Length of structure */
+	u32 Length;            /* Length of structure */
 	NDIS_802_11_MAC_ADDRESS Bssid;
-	ULONG Flags;
+	u32 Flags;
 } NDIS_802_11_AUTHENTICATION_REQUEST, *PNDIS_802_11_AUTHENTICATION_REQUEST;
 
 typedef enum _NDIS_802_11_STATUS_TYPE {
@@ -220,8 +220,8 @@ typedef struct _NDIS_802_11_AUTHENTICATION_EVENT {
 } NDIS_802_11_AUTHENTICATION_EVENT, *PNDIS_802_11_AUTHENTICATION_EVENT;
 
 typedef struct _NDIS_802_11_TEST {
-	ULONG Length;
-	ULONG Type;
+	u32 Length;
+	u32 Type;
 	union {
 		NDIS_802_11_AUTHENTICATION_EVENT AuthenticationEvent;
 		NDIS_802_11_RSSI RssiTrigger;
@@ -256,18 +256,18 @@ typedef struct _WLAN_BCN_INFO {
 *   WLAN_BSSID_EX and get_WLAN_BSSID_EX_sz()
 */
 typedef struct _WLAN_BSSID_EX {
-	ULONG  Length;
+	u32  Length;
 	NDIS_802_11_MAC_ADDRESS  MacAddress;
 	u8  Reserved[2];/* [0]: IS beacon frame */
 	NDIS_802_11_SSID  Ssid;
-	ULONG  Privacy;
+	u32  Privacy;
 	NDIS_802_11_RSSI  Rssi;/* (in dBM,raw data ,get from PHY) */
 	NDIS_802_11_NETWORK_TYPE  NetworkTypeInUse;
 	NDIS_802_11_CONFIGURATION  Configuration;
 	NDIS_802_11_NETWORK_INFRASTRUCTURE  InfrastructureMode;
 	NDIS_802_11_RATES_EX  SupportedRates;
 	WLAN_PHY_INFO	PhyInfo;
-	ULONG  IELength;
+	u32  IELength;
 	u8  IEs[MAX_IE_SZ];	/* (timestamp, beacon interval, and capability information) */
 }
 __attribute__((packed))
@@ -335,12 +335,12 @@ enum UAPSD_MAX_SP {
 
 typedef struct _PMKID_CANDIDATE {
 	NDIS_802_11_MAC_ADDRESS BSSID;
-	ULONG Flags;
+	u32 Flags;
 } PMKID_CANDIDATE, *PPMKID_CANDIDATE;
 
 typedef struct _NDIS_802_11_PMKID_CANDIDATE_LIST {
-	ULONG Version;       /* Version of the structure */
-	ULONG NumCandidates; /* No. of pmkid candidates */
+	u32 Version;       /* Version of the structure */
+	u32 NumCandidates; /* No. of pmkid candidates */
 	PMKID_CANDIDATE CandidateList[1];
 } NDIS_802_11_PMKID_CANDIDATE_LIST, *PNDIS_802_11_PMKID_CANDIDATE_LIST;
 
@@ -352,10 +352,10 @@ typedef struct _NDIS_802_11_AUTHENTICATION_ENCRYPTION {
 } NDIS_802_11_AUTHENTICATION_ENCRYPTION, *PNDIS_802_11_AUTHENTICATION_ENCRYPTION;
 
 typedef struct _NDIS_802_11_CAPABILITY {
-	ULONG  Length;
-	ULONG  Version;
-	ULONG  NoOfPMKIDs;
-	ULONG  NoOfAuthEncryptPairsSupported;
+	u32  Length;
+	u32  Version;
+	u32  NoOfPMKIDs;
+	u32  NoOfAuthEncryptPairsSupported;
 	NDIS_802_11_AUTHENTICATION_ENCRYPTION AuthenticationEncryptionSupported[1];
 
 } NDIS_802_11_CAPABILITY, *PNDIS_802_11_CAPABILITY;
